changed from template to inheritance in livingobject.js
character.js and weapon.js now inherit livingobject. i think the
inheritance metaphor is more understandable than template (as in C++).
but afterall it is only a change of concept, the underlying semantic
doesnt change, it is still prototypal javascript.

in addition I also changed some code in character.js related to fall down,
to fix a fall down bug. the corresponding constants in global.js are
changed.